+++ pg-ephemeral/src/lib.rs
@@ -4,6 +4,7 @@ pub mod config;
 pub mod container;
 pub mod definition;
 pub mod image;
+pub mod label;
 pub mod seed;
 
 pub use config::{Config, Instance};